+++ openvpn/ip/udp.hpp
@@ -52,19 +52,19 @@ inline std::uint16_t udp_checksum(const std::uint8_t *buf,
     /* calculate the sum of all 16 bit words */
     for (unsigned int i = 0; i < len_udp; i += 2)
     {
-        std::uint16_t word16 = ((buf[i] << 8) & 0xFF00) + ((i + 1 < len_udp) ? (buf[i + 1] & 0xFF) : 0);
+        std::uint16_t word16 = static_cast<uint16_t>(((buf[i] << 8) & 0xFF00) + ((i + 1 < len_udp) ? (buf[i + 1] & 0xFF) : 0));
         sum += word16;
     }
 
     /* add the UDP pseudo header which contains the IP source and destination addresses */
     for (unsigned int i = 0; i < 4; i += 2)
     {
-        std::uint16_t word16 = ((src_addr[i] << 8) & 0xFF00) + (src_addr[i + 1] & 0xFF);
+        std::uint16_t word16 = static_cast<uint16_t>(((src_addr[i] << 8) & 0xFF00) + (src_addr[i + 1] & 0xFF));
         sum += word16;
     }
     for (unsigned int i = 0; i < 4; i += 2)
     {
-        std::uint16_t word16 = ((dest_addr[i] << 8) & 0xFF00) + (dest_addr[i + 1] & 0xFF);
+        std::uint16_t word16 = static_cast<uint16_t>(((dest_addr[i] << 8) & 0xFF00) + (dest_addr[i + 1] & 0xFF));
         sum += word16;
     }
