+++ openvpn/openssl/sign/verify.hpp
@@ -30,6 +30,7 @@
 #include <openssl/bio.h>
 
 #include <openvpn/common/cleanup.hpp>
+#include <openvpn/common/numeric_cast.hpp>
 #include <openvpn/common/base64.hpp>
 #include <openvpn/openssl/pki/x509.hpp>
 #include <openvpn/openssl/util/error.hpp>
@@ -88,7 +89,7 @@ inline void verify(const OpenSSLPKI::X509 &cert,
     // verify signature
     EVP_VerifyInit(md_ctx, dig);
     EVP_VerifyUpdate(md_ctx, data.c_str(), data.length());
-    if (EVP_VerifyFinal(md_ctx, binsig.c_data(), binsig.length(), pkey) != 1)
+    if (EVP_VerifyFinal(md_ctx, binsig.c_data(), numeric_cast<unsigned int>(binsig.length()), pkey) != 1)
         throw OpenSSLException("OpenSSLSign::verify: verification failed");
 }
 } // namespace OpenSSLSign
