+++ openvpn/transport/client/httpcli.hpp
@@ -323,7 +323,7 @@ class Client : public TransportClient, AsyncResolvableTCP
     {
     }
 
-    unsigned int transport_send_queue_size() override
+    size_t transport_send_queue_size() override
     {
         if (impl)
             return impl->send_queue_size();
