+++ openvpn/transport/pktstream.hpp
@@ -26,6 +26,7 @@
 #include <limits>
 
 #include <openvpn/common/exception.hpp>
+#include <openvpn/common/numeric_cast.hpp>
 #include <openvpn/buffer/buffer.hpp>
 #include <openvpn/frame/frame.hpp>
 
@@ -124,7 +125,7 @@ class PacketStream
     // prepend uint16_t size to buffer
     static void prepend_size(Buffer &buf)
     {
-        const std::uint16_t net_len = htons(buf.size());
+        const std::uint16_t net_len = htons(numeric_cast<uint16_t>(buf.size()));
         buf.prepend((const unsigned char *)&net_len, sizeof(net_len));
     }
